chore: split up rendering code
This is done so that there is less risk of accidentally introducing a
race condition due to global variables
